eval/pyproject.toml
[project]
name = "eval"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "lpips>=0.1.4",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "scikit-image>=0.26.0",
    "torch>=2.10.0",
    "torch-fidelity>=0.4.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.2.0",
]
